+++ src/Websites/Websites/Cmdlets/AppServicePlans/NewAzureAppServicePlan.cs
@@ -18,6 +18,7 @@ using Microsoft.Azure.Commands.WebApps.Models.WebApp;
 using Microsoft.Azure.Commands.WebApps.Utilities;
 using Microsoft.Azure.Management.WebSites.Models;
 using System;
+using System.Collections.Generic;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.WebApps.Cmdlets.AppServicePlans
@@ -63,6 +64,9 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.AppServicePlans
         [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
         public SwitchParameter AsJob { get; set; }
 
+        [Parameter(ParameterSetName = ParameterSet1Name, Position = 7, Mandatory = false, HelpMessage = "Tags are name/value pairs that enable you to categorize resources")]
+        public IDictionary<string, string> Tags { get; set; }
+
         public override void ExecuteCmdlet()
         {
             if (HyperV.IsPresent && Tier != "PremiumContainer")
@@ -107,7 +111,8 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.AppServicePlans
                 Location = Location,
                 Sku = sku,
                 PerSiteScaling = PerSiteScaling,
-                IsXenon = HyperV.IsPresent
+                IsXenon = HyperV.IsPresent,
+                Tags = Tags
             };
 
             AppServicePlan retPlan = WebsitesClient.CreateOrUpdateAppServicePlan(ResourceGroupName, Name, appServicePlan, AseName, aseResourceGroupName);