+++ src/Websites/Websites/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Adding Tag parameter for New-AzAppServicePlan and Set-AzAppServicePlan &
+ Stop executing if an exception is thrown when adding a custom domain to a website.
 * Set-AzWebapp and Set-AzWebappSlot supports AlwaysOn, MinTls and FtpsState properties
 * Fixing issue where setting HttpsOnly along with changing AppservicePlan at the same time using the single Set-AzWebApp Command, was resetting HttpsOnly to default value